reordered the actions, so clang-format and doxygen are run first
This change reorders the actions so the least expensive and most failing
actions are run first. On a typical run, doxygen and clang-format take
each about 2s (plus 7s install), the complete building on all the supported platforms
10 minutes. In my experience, the most fails I get are from formatting
and documentation errors, not from compilation problems. This gives the
users a much faster feedback on needed changes in their commits and PRs
and shaves off a lot of execution time.